Release 2.11.0 with Android Headless JS support and more
