--- index.ios.js
@@ -1,3 +1,0 @@
-import API from "./RNBackgroundGeolocation";
-
-module.exports = API;
