+++ ios/RNBackgroundGeolocation/TSLocationManager.framework/_CodeSignature/CodeResources
@@ -54,7 +54,7 @@
 		</data>
 		<key>Info.plist</key>
 		<data>
-		n6f3Vuubsk9p4U6capHfGEMngII=
+		mBf8Vy5to3Gb3vTBrscMplSQW2E=
 		</data>
 		<key>Modules/module.modulemap</key>
 		<data>
